Item Pickup Drop and Follow Mechanic.
Scriptable Objects/Pickups.meta
> New Pickups folder for Scriptable Objects.
EXP.asset
> New EXP Scriptable Object.
EventManager.cs
> Added a new OnExpDrop event which entails Vector3.
ObjectPooler.cs
> Added a new expSo (PickupScriptableObject) variable.
PickupManager.cs
> Registers to SingletonManager.
> Gets objectPooler reference.
> Listens to onExpDrop invokes. (lines 25 to 34).
> SpawnsPickup with parameter Vector3.
SpawnManager.cs
> Formerly known as EnemySpawner.
Scripts/Pickups.meta
> New Pickups folder.
Health_Pickup.cs
> Inherits from Pickup.cs.
> Heals player on Item Pickup.
Pickup.cs
> Base pickup script.
> Gets rb (Rigidbody2D) reference.
> Follows Player position once SetTarget() has commenced.
> Has variable pickupSpeed (float); which can be edited through Unity Inspector.
> Has variable amount (float); which can be edited through Unity Inspector.
PickupScriptableObject.cs
> New Pickup Scriptable Object script.
> Inherits from PoolScriptableObject.cs
XP_Pickup.cs
> Inherits from Pickup.cs
> Adds XP on pickup.
EnemyStat.cs
> Now drops EXP instead of granting Player EXP "on kill" (line 11).
Health.cs
> Heal() parameter changed to float; from int.
Magnet.cs
> New Magnet script.
> Grants Player position through parenting. (The script component is a child of the Player)
PlayerStat.cs
> Added overrides just for Player to use.
Stat.cs
> Changed Heal() parameter to float; from int.

+++ Assets/Scripts/Units/Health.cs
@@ -46,7 +46,7 @@ public class Health
         // Debug.Log("Death() is called.");
     }
     
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         if (health < maxHealth)
         {
